Refractor code for better readability Add support of repeated hit object
